renamed files based on feedback. added logging to avoid simplying doing a pass
