+++ cloudcomposerdiff/src/cloudcomposerdiff/__main__.py
@@ -19,8 +19,8 @@ from google.cloud.orchestration.airflow import service_v1
 
 from cloudcomposerdiff.lib.comparator import EnvironmentComparator
 from cloudcomposerdiff.lib.service import GCPComposerService
-from cloudcomposerdiff.lib.strategies.diffairflowconfig import DiffAirflowConfig
-from cloudcomposerdiff.lib.strategies.diffenvimage import DiffEnvImage
+from cloudcomposerdiff.lib.strategies.diff_airflow_config import DiffAirflowConfig
+from cloudcomposerdiff.lib.strategies.diff_env_image import DiffEnvImage
 
 from . import __version__
