+++ cloudcomposerdiff/src/cloudcomposerdiff/lib/strategies/diff_env_image.py
